tidy up + reformat
