+++ cmake/modules/IkarusMacros.cmake
@@ -20,10 +20,9 @@ include(AddSpectraFlags)
 
 set(MUESLI_ROOT "/muesli")
 add_library(muesli STATIC IMPORTED)
-set_target_properties(muesli PROPERTIES
-    IMPORTED_LOCATION "${MUESLI_ROOT}/lib/libmuesli_linux_ia64.a"
-    INTERFACE_INCLUDE_DIRECTORIES "${MUESLI_ROOT}"
+set_target_properties(
+  muesli PROPERTIES IMPORTED_LOCATION "${MUESLI_ROOT}/lib/libmuesli_linux_ia64.a"
+                    INTERFACE_INCLUDE_DIRECTORIES "${MUESLI_ROOT}"
 )
 
 include(AddMuesliFlags)
-
