+++ ikarus/finiteelements/mechanics/materials/muesli/mueslifinite.hh
@@ -4,7 +4,7 @@
 /**
  * \file Muesli.hh
  * \brief Implementation of the Muesli material model.
- * \ingroup  materials
+ * \ingroup materials
  */
 
 #pragma once
@@ -19,19 +19,19 @@
 
 namespace Ikarus::Materials::Muesli {
 
-template <typename SM = muesli::elasticIsotropicMaterial>
-requires(std::is_base_of_v<muesli::smallStrainMaterial, SM>)
-struct SmallStrain : public Material<SmallStrain<SM>>
+template <typename FM>
+requires(std::is_base_of_v<muesli::finiteStrainMaterial, FM>)
+struct FiniteStrain : public Material<FiniteStrain<FM>>
 {
-  using MaterialModel                 = SM;
+  using MaterialModel                 = FM;
   using ScalarType                    = double;
   static constexpr int worldDimension = 3;
   using StrainMatrix                  = Eigen::Matrix<ScalarType, worldDimension, worldDimension>;
   using StressMatrix                  = StrainMatrix;
   using MaterialParameters            = Muesli::MaterialProperties;
 
-  static constexpr auto strainTag              = StrainTags::linear;
-  static constexpr auto stressTag              = StressTags::linear;
+  static constexpr auto strainTag              = StrainTags::rightCauchyGreenTensor;
+  static constexpr auto stressTag              = StressTags::PK2;
   static constexpr auto tangentModuliTag       = TangentModuliTags::Material;
   static constexpr bool energyAcceptsVoigt     = false;
   static constexpr bool stressToVoigt          = false;
@@ -40,24 +40,15 @@ struct SmallStrain : public Material<SmallStrain<SM>>
   static constexpr bool moduliAcceptsVoigt     = false;
   static constexpr double derivativeFactorImpl = 1;
 
-  [[nodiscard]] constexpr static std::string nameImpl() noexcept {
-    return "Muesli_SmallStrain: " + Dune::className<SM>();
-  }
+  [[nodiscard]] constexpr static std::string nameImpl() noexcept { return "FiniteStrain: " + materialName<FM>(); }
 
   /**
-   * \brief Constructor for MuesliT.
-   * \param mpt The Lame's parameters (first parameter and shear modulus).
+   * \briefCConstructor for FiniteStrain muesli materials
+   * \param mpt Muesli materialproperties
    */
-  template <typename MPT>
-  requires(std::same_as<MPT, YoungsModulusAndPoissonsRatio> or std::same_as<MPT, LamesFirstParameterAndShearModulus>)
-  explicit SmallStrain(const MPT& mpt)
-      : materialParameter_{Muesli::propertiesFromIkarusMaterialParameters(mpt)},
-        material_{Dune::className<SM>(), materialParameter_},
-        mp_{material_.createMaterialPoint()} {}
-
-  explicit SmallStrain(const MaterialParameters& mpt)
+  explicit FiniteStrain(const MaterialParameters& mpt)
       : materialParameter_{mpt},
-        material_{Dune::className<SM>(), mpt},
+        material_{Dune::className<FM>(), mpt},
         mp_{material_.createMaterialPoint()} {}
 
   /**
@@ -72,21 +63,22 @@ struct SmallStrain : public Material<SmallStrain<SM>>
    * \return ScalarType The stored energy.
    */
   template <typename Derived>
-  ScalarType storedEnergyImpl(const Eigen::MatrixBase<Derived>& E) const {
+  ScalarType storedEnergyImpl(const Eigen::MatrixBase<Derived>& C) const {
     static_assert(Concepts::EigenMatrixOrVoigtNotation3<Derived>);
     if constexpr (!Concepts::EigenVector<Derived>) {
-      updateState(E);
+      updateState(C);
       return mp_->storedEnergy();
+
     } else
       static_assert(!Concepts::EigenVector<Derived>,
-                    "Muesli energy can only be called with a matrix and not a vector in Voigt notation");
+                    "MuesliFiniteStrain energy can only be called with a matrix and not a vector in Voigt notation");
   }
 
   /**
    * \brief Computes the stresses in the Neo-Hookean material model.
    * \tparam voigt A boolean indicating whether to return stresses in Voigt notation.
    * \tparam Derived The derived type of the input matrix.
-   * \param E The right Cauchy-Green tensor.
+   * \param C The right Cauchy-Green tensor.
    * \return StressMatrix The stresses.
    */
   template <bool voigt, typename Derived>
@@ -95,60 +87,68 @@ struct SmallStrain : public Material<SmallStrain<SM>>
     if constexpr (!voigt) {
       if constexpr (!Concepts::EigenVector<Derived>) {
         updateState(C);
-        mp_->stress(stress_);
+        mp_->secondPiolaKirchhoffStress(stress_);
         return Muesli::toMatrix<ScalarType>(stress_);
       } else
         static_assert(!Concepts::EigenVector<Derived>,
-                      "Muesli can only be called with a matrix and not a vector in Voigt notation");
+                      "MuesliFiniteStrain can only be called with a matrix and not a vector in Voigt notation");
     } else
-      static_assert(voigt == false, "Muesli does not support returning stresses in Voigt notation");
+      static_assert(voigt == false, "MuesliFiniteStrain does not support returning stresses in Voigt notation");
   }
 
   /**
    * \brief Computes the tangent moduli in the Neo-Hookean material model.
    * \tparam voigt A boolean indicating whether to return tangent moduli in Voigt notation.
    * \tparam Derived The derived type of the input matrix.
-   * \param E The right Cauchy-Green tensor.
+   * \param C The right Cauchy-Green tensor.
    * \return Eigen::TensorFixedSize<ScalarType, Eigen::Sizes<3, 3, 3, 3>> The tangent moduli.
    */
   template <bool voigt, typename Derived>
-  auto tangentModuliImpl(const Eigen::MatrixBase<Derived>& E) const {
+  auto tangentModuliImpl(const Eigen::MatrixBase<Derived>& C) const {
     static_assert(Concepts::EigenMatrixOrVoigtNotation3<Derived>);
     if constexpr (!voigt) {
       if constexpr (!Concepts::EigenVector<Derived>) {
-        updateState(E);
-        mp_->tangentTensor(tangentModuli_);
+        updateState(C);
+
+        mp_->convectedTangent(tangentModuli_);
         return Muesli::toTensor<ScalarType>(tangentModuli_);
       } else
         static_assert(!Concepts::EigenVector<Derived>,
-                      "Muesli can only be called with a matrix and not a vector in Voigt notation");
+                      "MuesliFiniteStrain can only be called with a matrix and not a vector in Voigt notation");
     } else
-      static_assert(voigt == false, "Muesli does not support returning tangent moduli in Voigt notation");
+      static_assert(voigt == false, "MuesliFiniteStrain does not support returning tangent moduli in Voigt notation");
   }
 
+  /**
+   * \brief Returns the underlying muesli material implementation
+   * \return auto& reference to the musli material
+   */
   auto& material() const { return material_; }
 
+  /**
+   * \brief asserts that the materialpoint pointer is not null
+   */
   bool assertMP() const { return mp_.get() != NULL; }
 
-  SmallStrain(const SmallStrain& other)
+  FiniteStrain(const FiniteStrain& other)
       : materialParameter_{other.materialParameter_},
-        material_{Dune::className<SM>(), materialParameter_},
+        material_{Dune::className<FM>(), materialParameter_},
         mp_{material_.createMaterialPoint()} {}
 
 private:
   MaterialParameters materialParameter_;
   MaterialModel material_;
-  std::unique_ptr<muesli::smallStrainMP> mp_;
+  std::unique_ptr<muesli::finiteStrainMP> mp_;
 
   mutable istensor strain_{};
   mutable istensor stress_{};
   mutable itensor4 tangentModuli_{};
 
   template <typename Derived>
-  void updateState(const Eigen::MatrixBase<Derived>& E) const {
-    Muesli::toistensor(strain_, E);
+  void updateState(const Eigen::MatrixBase<Derived>& C) const {
+    Muesli::toistensor(strain_, transformStrain<strainTag, StrainTags::deformationGradient>(C));
     mp_->updateCurrentState(0.0, strain_);
   }
 };
 
-} // namespace Ikarus::Materials
+} // namespace Ikarus::Materials::Muesli
