+++ ikarus/finiteelements/mechanics/materials/muesli/mueslismall.hh
@@ -4,7 +4,7 @@
 /**
  * \file Muesli.hh
  * \brief Implementation of the Muesli material model.
- * \ingroup  materials
+ * \ingroup materials
  */
 
 #pragma once
@@ -19,19 +19,19 @@
 
 namespace Ikarus::Materials::Muesli {
 
-template <typename FM>
-requires(std::is_base_of_v<muesli::finiteStrainMaterial, FM>)
-struct FiniteStrain : public Material<FiniteStrain<FM>>
+template <typename SM = muesli::elasticIsotropicMaterial>
+requires(std::is_base_of_v<muesli::smallStrainMaterial, SM>)
+struct SmallStrain : public Material<SmallStrain<SM>>
 {
-  using MaterialModel                 = FM;
+  using MaterialModel                 = SM;
   using ScalarType                    = double;
   static constexpr int worldDimension = 3;
   using StrainMatrix                  = Eigen::Matrix<ScalarType, worldDimension, worldDimension>;
   using StressMatrix                  = StrainMatrix;
   using MaterialParameters            = Muesli::MaterialProperties;
 
-  static constexpr auto strainTag              = StrainTags::rightCauchyGreenTensor;
-  static constexpr auto stressTag              = StressTags::PK2;
+  static constexpr auto strainTag              = StrainTags::linear;
+  static constexpr auto stressTag              = StressTags::linear;
   static constexpr auto tangentModuliTag       = TangentModuliTags::Material;
   static constexpr bool energyAcceptsVoigt     = false;
   static constexpr bool stressToVoigt          = false;
@@ -40,17 +40,26 @@ struct FiniteStrain : public Material<FiniteStrain<FM>>
   static constexpr bool moduliAcceptsVoigt     = false;
   static constexpr double derivativeFactorImpl = 1;
 
-  [[nodiscard]] constexpr static std::string nameImpl() noexcept {
-    return "Muesli_FiniteStrain: " + Dune::className<FM>();
-  }
+  [[nodiscard]] constexpr static std::string nameImpl() noexcept { return "Muesli_SmallStrain: " + materialName<SM>(); }
 
   /**
-   * \brief Constructor for MuesliT.
+   * \brief Constructor for SmallStrain muesli materials (only activated for isotropic linear elasticity).
    * \param mpt The Lame's parameters (first parameter and shear modulus).
    */
-  explicit FiniteStrain(const MaterialParameters& mpt)
+  template <typename MPT>
+  requires(std::same_as<MaterialModel, muesli::elasticIsotropicMaterial>)
+  explicit SmallStrain(const MPT& mpt)
+      : materialParameter_{Muesli::propertiesFromIkarusMaterialParameters(mpt)},
+        material_{Dune::className<SM>(), materialParameter_},
+        mp_{material_.createMaterialPoint()} {}
+
+  /**
+   * \brief CConstructor for SmallStrain muesli materials
+   * \param mpt Muesli materialproperties
+   */
+  explicit SmallStrain(const MaterialParameters& mpt)
       : materialParameter_{mpt},
-        material_{Dune::className<FM>(), mpt},
+        material_{Dune::className<SM>(), mpt},
         mp_{material_.createMaterialPoint()} {}
 
   /**
@@ -65,22 +74,21 @@ struct FiniteStrain : public Material<FiniteStrain<FM>>
    * \return ScalarType The stored energy.
    */
   template <typename Derived>
-  ScalarType storedEnergyImpl(const Eigen::MatrixBase<Derived>& C) const {
+  ScalarType storedEnergyImpl(const Eigen::MatrixBase<Derived>& E) const {
     static_assert(Concepts::EigenMatrixOrVoigtNotation3<Derived>);
     if constexpr (!Concepts::EigenVector<Derived>) {
-      updateState(C);
+      updateState(E);
       return mp_->storedEnergy();
-
     } else
       static_assert(!Concepts::EigenVector<Derived>,
-                    "Muesli energy can only be called with a matrix and not a vector in Voigt notation");
+                    "MuesliSmallStrain energy can only be called with a matrix and not a vector in Voigt notation");
   }
 
   /**
    * \brief Computes the stresses in the Neo-Hookean material model.
    * \tparam voigt A boolean indicating whether to return stresses in Voigt notation.
    * \tparam Derived The derived type of the input matrix.
-   * \param C The right Cauchy-Green tensor.
+   * \param E The liinear strain tensor.
    * \return StressMatrix The stresses.
    */
   template <bool voigt, typename Derived>
@@ -89,61 +97,67 @@ struct FiniteStrain : public Material<FiniteStrain<FM>>
     if constexpr (!voigt) {
       if constexpr (!Concepts::EigenVector<Derived>) {
         updateState(C);
-        mp_->secondPiolaKirchhoffStress(stress_);
+        mp_->stress(stress_);
         return Muesli::toMatrix<ScalarType>(stress_);
       } else
         static_assert(!Concepts::EigenVector<Derived>,
-                      "Muesli can only be called with a matrix and not a vector in Voigt notation");
+                      "MuesliSmallStrain can only be called with a matrix and not a vector in Voigt notation");
     } else
-      static_assert(voigt == false, "Muesli does not support returning stresses in Voigt notation");
+      static_assert(voigt == false, "MuesliSmallStrain does not support returning stresses in Voigt notation");
   }
 
   /**
    * \brief Computes the tangent moduli in the Neo-Hookean material model.
    * \tparam voigt A boolean indicating whether to return tangent moduli in Voigt notation.
    * \tparam Derived The derived type of the input matrix.
-   * \param C The right Cauchy-Green tensor.
+   * \param E The liinear strain tensor.
    * \return Eigen::TensorFixedSize<ScalarType, Eigen::Sizes<3, 3, 3, 3>> The tangent moduli.
    */
   template <bool voigt, typename Derived>
-  auto tangentModuliImpl(const Eigen::MatrixBase<Derived>& C) const {
+  auto tangentModuliImpl(const Eigen::MatrixBase<Derived>& E) const {
     static_assert(Concepts::EigenMatrixOrVoigtNotation3<Derived>);
     if constexpr (!voigt) {
       if constexpr (!Concepts::EigenVector<Derived>) {
-        updateState(C);
-
-        mp_->convectedTangent(tangentModuli_);
+        updateState(E);
+        mp_->tangentTensor(tangentModuli_);
         return Muesli::toTensor<ScalarType>(tangentModuli_);
       } else
         static_assert(!Concepts::EigenVector<Derived>,
-                      "Muesli can only be called with a matrix and not a vector in Voigt notation");
+                      "MuesliSmallStrain can only be called with a matrix and not a vector in Voigt notation");
     } else
-      static_assert(voigt == false, "Muesli does not support returning tangent moduli in Voigt notation");
+      static_assert(voigt == false, "MuesliSmallStrain does not support returning tangent moduli in Voigt notation");
   }
 
+  /**
+   * \brief Returns the underlying muesli material implementation
+   * \return auto& reference to the musli material
+   */
   auto& material() const { return material_; }
 
+  /**
+   * \brief asserts that the materialpoint pointer is not null
+   */
   bool assertMP() const { return mp_.get() != NULL; }
 
-  FiniteStrain(const FiniteStrain& other)
+  SmallStrain(const SmallStrain& other)
       : materialParameter_{other.materialParameter_},
-        material_{Dune::className<FM>(), materialParameter_},
+        material_{Dune::className<SM>(), materialParameter_},
         mp_{material_.createMaterialPoint()} {}
 
 private:
   MaterialParameters materialParameter_;
   MaterialModel material_;
-  std::unique_ptr<muesli::finiteStrainMP> mp_;
+  std::unique_ptr<muesli::smallStrainMP> mp_;
 
   mutable istensor strain_{};
   mutable istensor stress_{};
   mutable itensor4 tangentModuli_{};
 
   template <typename Derived>
-  void updateState(const Eigen::MatrixBase<Derived>& C) const {
-    Muesli::toistensor(strain_, transformStrain<strainTag, StrainTags::deformationGradient>(C));
+  void updateState(const Eigen::MatrixBase<Derived>& E) const {
+    Muesli::toistensor(strain_, E);
     mp_->updateCurrentState(0.0, strain_);
   }
 };
 
-} // namespace Ikarus::Materials
+} // namespace Ikarus::Materials::Muesli
